this cover of gerudo valley that i'm listening to would make a great bit of soundtrack for like the opening scene of a zelda movie
like i can almost storyboard it in my head. some sort of desert chase scene, introducing ganondorf. dunno, would you want to have the link/ganondorf be a strong interplay in a zelda movie? probably would actually want to focus on the link/zelda dynamic more, in which case starting with the villain might not be the right call